requirements.txt
langchain
langchain-community
langgraph
pypdf
pymupdf
faiss-cpu
sentence-transformers
python-dotenv
openai
langchain-openai
langchain-huggingface
